debit card in DB is renamed on fakeDebitCards and added new columns in it, months enums are added, booking isPaid column added and payment method is almost ready

+++ RentalCars/RentalCars.Core/Services/DebitCards/IDebitCardService.cs
@@ -2,6 +2,6 @@
 {
     public interface IDebitCardService
     {
-        int CreateDebitCard(int creditCardNumber, int cvv, string fullNameOnCard, string expMonth, int expYear);
+        int CreateDebitCard(long creditCardNumber, int cvv, string fullNameOnCard, string expMonth, int expYear);
     }
 }
